% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adj.risk}
\alias{adj.risk}
\title{Adjusted risk}
\usage{
adj.risk(rr, ppr)
}
\arguments{
\item{rr}{relative risk values (vector)}

\item{ppr}{population proportions corresponding to
rr values (vector of equal length to rr)}
}
\value{
vector of adjusted risk values (in order corresponding to rr)
}
\description{
Calculates adjusted risk for given
  relative risk and population proportions
}
\examples{
# examples for adj.risk
adj.risk(c(5, 1), c(0.1, 0.9))
adj.risk(c(5, 3, 1), c(0.1, 0.1, 0.8))
}
\keyword{methods}

