% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sd.tp}
\alias{sd.tp}
\title{Standard deviation of true prevalence estimate}
\usage{
sd.tp(x, n, se, sp)
}
\arguments{
\item{x}{number of positive results in sample}

\item{n}{sample size}

\item{se}{test sensitivity}

\item{sp}{test specificity}
}
\value{
vector of standard deviation values for true prevalence estimates
}
\description{
Calculates the standard deviation of true prevalence estimate
assuming se and sp known exactly, used to calculate normal approximation CI for estimate
}
\examples{
# example of sd.tp
sd.tp(1:10, 20, 0.9, 0.99)
}
\keyword{methods}

