% Generated by roxygen2 (4.0.2): do not edit by hand
\name{se.parallel}
\alias{se.parallel}
\title{Sensitivity of tests in parallel}
\usage{
se.parallel(se)
}
\arguments{
\item{se}{vector of unit sensitivity values}
}
\value{
scalar of combined sensitivity, assuming independence
}
\description{
Calculates the combined sensitivity for multiple tests
  interpreted in parallel (assuming independence)
}
\examples{
# examples for se.parallel
se.parallel(c(0.99, 0.95, 0.8))
}
\keyword{methods}

