% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sep.binom.imperfect}
\alias{sep.binom.imperfect}
\title{Binomial population sensitivity for imperfect test}
\usage{
sep.binom.imperfect(n, c = 1, se, sp = 1, pstar)
}
\arguments{
\item{n}{sample size}

\item{c}{The cut-point number of reactors to classify a herd/flock
as positive, default=1, if reactors < c result is negative, >= c is positive}

\item{se}{test unit sensitivity}

\item{sp}{test unit specificity, default=1}

\item{pstar}{design prevalence}
}
\value{
a vector of population-level sensitivities
}
\description{
Calculates population sensitivity for a large or unknown population
  and allowing for imperfect test sensitivity and specificity,
  using Binomial distribution
}
\examples{
# examples for sep.imperfect.binom
sep.binom.imperfect(1:10*5, 2, 0.95, 0.98, 0.1)
sep.binom.imperfect(50, 1:5, 0.95, 0.98, 0.1)
sep.binom.imperfect(30, 2, 0.9, 0.98, 0.1)
sep.binom.imperfect(30, 1, 0.9, 0.98, 0.1)
}
\keyword{methods}

