% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sep.binom}
\alias{sep.binom}
\title{Binomial Population sensitivity}
\usage{
sep.binom(n, pstar, se = 1, sp = 1)
}
\arguments{
\item{n}{integer scalar or vector of number tested (sample size)}

\item{pstar}{scalar or vector of design prevalence as proportion}

\item{se}{unit sensitivity of test (proportion), default = 1}

\item{sp}{unit specificity of test (proportion), default = 1}
}
\value{
vector of population-level sensitivities
}
\description{
Calculates population sensitivity for detecting disease,
  assuming imperfect test sensitivity and specificity and representative sampling,
  using binomial distribution (assumes large or unknown population size)
}
\examples{
# examples for sep.binom - checked
sep.binom(n=300, pstar = 0.02, se = 0.92)
tested<- seq(10,100, by=10)
prev<- 0.05
sens<- 0.9
sep.binom(tested, prev, sens)
}
\keyword{methods}

