% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sep.hp}
\alias{sep.hp}
\title{Hypergeometric (HerdPlus) population sensitivity for imperfect test}
\usage{
sep.hp(N, n, c = 1, se, sp = 1, pstar)
}
\arguments{
\item{N}{population size}

\item{n}{sample size}

\item{c}{The cut-point number of reactors to classify a herd/flock
as positive, default=1, if reactors < c result is negative, >= c is positive}

\item{se}{test unit sensitivity}

\item{sp}{test unit specificity, default=1}

\item{pstar}{design prevalence}
}
\value{
a vector of population-level sensitivities
}
\description{
Calculates population sensitivity for a finite population
  and allowing for imperfect test sensitivity and specificity,
  using Hypergeometric distribution
}
\examples{
# examples of sep.hp
sep.hp(150, 1:5*10, 2, 0.9, 0.98, 0.1)
sep.hp(150, 30, 2, 0.9, 0.98, 15)
sep.hp(150, 30, 1, 0.9, 0.98, 15)
sep.hp(150, 30, 1, 0.9, 0.98, 0.1)
}
\keyword{methods}

