% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sep.pfree}
\alias{sep.pfree}
\title{Population sensitivity to achieve desired (posterior) probability of freedom}
\usage{
sep.pfree(prior, pfree)
}
\arguments{
\item{prior}{prior probability of freedom before surveillance (scalar or vector)}

\item{pfree}{desired probability of freedom (scalar or vector)}
}
\value{
a vector of population-level sensitivities
}
\description{
Calculates the population sensitivity required to achieve a
  given value for probability of disease freedom
}
\examples{
# examples of sep.pfree
sep.pfree(0.5, 0.95)
sep.pfree(c(0.5, 0.6, 0.7, 0.8, 0.9, 0.95), 0.99)
sep.pfree(0.5, c(0.8, 0.9, 0.95, 0.99))
}
\keyword{methods}

