% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedom_functions_1.R
\name{sep.prior}
\alias{sep.prior}
\title{Population sensitivity to achieve desired prior probability of freedom}
\usage{
sep.prior(prior, p.intro)
}
\arguments{
\item{prior}{prior probability of freedom before surveillance (scalar or vector)}

\item{p.intro}{probability of introduction for time period (scalar or vector equal length to sep)}
}
\value{
a vector of population-level sensitivities
}
\description{
Calculates the population sensitivity required to achieve a
  given value for the prior (discounted) probability of disease freedom
}
\examples{
# examples of sep.prior
sep.prior(0.95, 0.01)
sep.prior(c(0.9, 0.95, 0.98, 0.99), 0.01)
sep.prior(0.95, c(0.001, 0.005, 0.01, 0.02, 0.05))
}
\keyword{methods}

