% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sp.series}
\alias{sp.series}
\title{Specficity of tests in series}
\usage{
sp.series(sp)
}
\arguments{
\item{sp}{vector of unit specificity values}
}
\value{
scalar of combined specificity, assuming independence
}
\description{
Calculates the combined specificity for multiple tests
  interpreted in series (assuming independence)
}
\examples{
# examples for sp.series
sp.series(c(0.99, 0.95, 0.8))
}
\keyword{methods}

