% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sph.hp}
\alias{sph.hp}
\title{Hypergeometric population specificity calculation}
\usage{
sph.hp(N, n, c = 1, sp)
}
\arguments{
\item{N}{population size}

\item{n}{sample size}

\item{c}{The cut-point number of reactors to classify a herd/flock
as positive, default=1, if reactors < c result is negative, >= c is positive}

\item{sp}{test unit specificity}
}
\value{
a vector of population-level specificities
}
\description{
Calculates population specificity for a finite population
and imperfect test, using Hypergeometric distribution
}
\examples{
# examples of sph.hp
sph.hp(150, 30, 2, 0.98)
sph.hp(150, 30, 1, 0.98)
sph.hp(150, 1:5*10, 2, 0.98)
sph.hp(500, 30, 2, 95:100/100)
}
\keyword{methods}

