% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sph}
\alias{sph}
\title{Population specificity}
\usage{
sph(n, sp)
}
\arguments{
\item{n}{integer scalar or vector of number tested}

\item{sp}{unit specificity of test (proportion)}
}
\value{
a vector of population-level specificities
}
\description{
Calculates population specificity assuming representative sampling
}
\examples{
# examples for sph - checked
sph(10, 0.9)
sph(c(10, 20, 50, 100), 0.99)
sph(100, c(0.999, 0.99, 0.98, 0.95, 0.9))
}
\keyword{methods}

