% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freedom_functions_1.R
\name{spp}
\alias{spp}
\title{Population specificity}
\usage{
spp(n, sp)
}
\arguments{
\item{n}{sample size (number tested), integer, scalar or vector}

\item{sp}{unit specificity of test (proportion), scalar or vector of same length as n}
}
\value{
a vector of population-level specificities
}
\description{
Calculates population specificity assuming representative sampling
}
\examples{
# examples for spp - checked
spp(10, 0.9)
spp(c(10, 20, 50, 100), 0.99)
spp(100, c(0.999, 0.99, 0.98, 0.95, 0.9))
}
\keyword{methods}

