% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_based_functions.R
\name{adj.risk}
\alias{adj.risk}
\title{Adjusted risk}
\usage{
adj.risk(rr, ppr)
}
\arguments{
\item{rr}{relative risk values (vector of values corresponding to the number of risk strata)}

\item{ppr}{population proportions corresponding to 
rr values (vector of equal length to rr)}
}
\value{
vector of adjusted risk values (in order corresponding to rr)
}
\description{
Calculates adjusted risk for given 
  relative risk and population proportions. This is an intermediate calculation
  in the calculation of effective probability of infection for risk-based 
  surveillance activities
}
\examples{
# examples for adj.risk
adj.risk(c(5, 1), c(0.1, 0.9))
adj.risk(c(5, 3, 1), c(0.1, 0.1, 0.8))
}
\keyword{methods}

