% Generated by roxygen2 (4.0.2): do not edit by hand
\name{n.pfree}
\alias{n.pfree}
\title{Sample size to achieve desired (posterior) probability of freedom}
\usage{
n.pfree(pfree, prior, p.intro, pstar, se, N = NA)
}
\arguments{
\item{pfree}{desired probability of freedom (scalar or vector)}

\item{prior}{prior probability of freedom before surveillance}

\item{p.intro}{probability of introduction for time period (scalar
or vector equal length to sep)}

\item{pstar}{design prevalence}

\item{se}{unit sensitivity (scalar or vector)}

\item{N}{population size}
}
\value{
vector of sample sizes
}
\description{
Calculates the sample size required to achieve a
  given value for probability of disease freedom
}
\examples{
# examples for n.pfree
n.pfree(0.95, 0.5, 0.01, 0.05, 0.9)
n.pfree(0.95, 0.5, 0.01, 0.05, 0.9, N=300)
n.pfree(pfree = c(0.9, 0.95, 0.98, 0.99), prior = 0.7, 0.01, 0.01, 0.8, 1000)
n.pfree(0.95, 0.7, 0.01, 0.1, 0.96)
}
\keyword{methods}

